pyproject.toml
[project]
name = "ligo-at"
version = "0.0.1"
description = "Decentralized links page on top of AT Protocol"
readme = "readme.md"
requires-python = ">=3.13"
dependencies = [
    "atproto>=0.0.62",
    "authlib>=1.3",
    "dnspython>=2.8.0",
    "flask[dotenv]>=3.1.2",
    "requests>=2.32",
    "requests-hardened>=1.0.0b3",
]
